updated Cmd_returnball and ItemBattleEffects

+++ include/battle_ai_script_commands.h
@@ -14,7 +14,7 @@ u8 BattleAI_ChooseMoveOrAction(void);
 void ClearBankMoveHistory(u8 bank);
 void RecordAbilityBattle(u32 bank, u32 abilityId);
 void ClearBankAbilityHistory(u8 bank);
-void RecordItemEffectBattle(u8 bank, u8 itemEffect);
+void RecordItemEffectBattle(u32 bank, u32 itemEffect);
 void ClearBankItemEffectHistory(u8 bank);
 u8 BattleAI_ChooseMoveOrAction(void);
